fluid: redo arg parsing
Ignore-this: b0ea3e534e85b021ae4f9b1847921857

darcs-hash:20120610071348-bf82d-998a532172f6ee0ee3a1e297a2bd3f52b97f8cf7

+++ gloss-examples/raster/Fluid/src-repa/UserEvent.hs
@@ -126,8 +126,8 @@ windowToModel config (x, y) = (x', y')
  where  (scaleX, scaleY)                = configScale config
         (windowWidth, windowHeight)     = configWindowSize config
 
-        x' = round ((x + (fromIntegral windowWidth  / 2)) / scaleX)
-        y' = round ((y + (fromIntegral windowHeight / 2)) / scaleY)
+        x' = round ((x + (fromIntegral windowWidth  / 2)) / fromIntegral scaleX)
+        y' = round ((y + (fromIntegral windowHeight / 2)) / fromIntegral scaleY)
